Get contemporaries from Elasticsearch
The benefits of this are twofold:

* it’s faster; and
* it ranks the people by the number of publications in shared subjects,
  which @frankieroberto & I think is likely to be a better metric.

The downside is that we can’t retrieve *all* people born around the same
time, since that number might be very large, and there’s a limit to how
many IDs we can pass into the subsequent query. So we’re just picking
the first 1000, which may or may not contain meaningful contemporaries.
We will revisit later if possible.

Closes #62.